wip: add pre-FF ui back and adjust FF
